Ignore unknown fields in bind_pids validation
Validare bind pids early, but validate only on the values we know
are expected to exist in the configuration. When bind pids executes
we don't immediately know the entity_type or desired_pid, these are
derived later-on. Let us focus on making sure the Handle server form
is filled in-correctly and ensure that the behavior of an exception
is known.