Dashboard: improve processing storage usage page
To avoid timeouts on the first load of this page, the usage data is only
calculated when the `calculate` get parameter is evaluated as `True`.
When the usage data is not calculated, the page shows a description and
a button to reload the page calculating the usage. When the usage is
calculated, the page will display a general information section and a
table of clearable directories from within the shared path.

- Show the root of the filesystem where the shared directory is located.
- Only show the size of the root path.
- Rename `_usage_dirs`function to`_get_shared_dirs` and change default
  `calculate_usage` to`False`.
- Simplify directory list generation and only include clearable dirs.
- Add`--one-file-system` option to `du` calls.
- Fix translations.
- Cosmetic changes.
- Add tests.